+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+* add retry behavior for nodes - recovery from errors
 
 ## [0.4.1] - 2022-01-23
 ### Changed